+++ frontend/src/Activity/ActTable.svelte
@@ -1,10 +1,11 @@
 <script lang="ts">
   import { parts, fmtNumber, fmtSeconds } from "../lib/store";
   import SvelteTable from "../Widgets/SvelteTable.svelte";
-  import { Activity, Usage } from "../lib/types";
   import RangeSlider from "svelte-range-slider-pips";
   import { Alert } from "@sveltestrap/sveltestrap";
   import { by } from "../lib/mapable";
+  import { Activity } from "./activity";
+  import { Usage } from "../Usage/usage";
 
   export let acts: Activity[];
