+++ frontend/src/Part/GearCard.svelte
@@ -8,8 +8,9 @@
     Col,
   } from "@sveltestrap/sveltestrap";
   import { types, usages, fmtSeconds, fmtDate, fmtNumber } from "../lib/store";
-  import { Part, Usage } from "../lib/types";
+  import { Part } from "../lib/types";
   import { link } from "svelte-spa-router";
+  import { Usage } from "../Usage/usage";
 
   export let part: Part;
   export let display = false;
