+++ frontend/src/Spares/Spares.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { filterValues } from "../lib/mapable";
   import { types, category } from "../lib/store";
-  import Usage from "../Usage.svelte";
+  import Usage from "../Usage/Usage.svelte";
   import SpareType from "./SpareType.svelte";
   import { Table } from "@sveltestrap/sveltestrap";
