+++ frontend/src/lib/store.ts
@@ -1,15 +1,9 @@
 import { writable, type Writable } from "svelte/store";
 import { mapable, mapObject } from "./mapable";
-import {
-  Activity,
-  Part,
-  Attachment,
-  Type,
-  type ActType,
-  type User,
-  Usage,
-} from "./types";
+import { Part, Attachment, Type, type ActType, type User } from "./types";
 import { Service, services } from "../Service/service";
+import { Activity } from "../Activity/activity";
+import { Usage } from "../Usage/usage";
 
 export function fmtDate(date: Date | undefined) {
   return date ? date.toLocaleDateString(navigator.language) : "never";
